feat(refactor): [Backend] Removed unnecessary retrieval of liked users from getPosts and added LikedUsers feature
In the backend, the getPosts function of the posts controller was refactored to remove the unnecessary retrieval of the list of liked users for each post. Additionally, a new feature called LikedUsers was added to the controller, which retrieves the list of liked users separately. This improves code clarity and efficiency.